// src/api/models/HeadToHeadGeneralRecord.ts
/* tslint:disable */
/* eslint-disable */
/**
 * radarvan
 * Stats for generals
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { General } from './General';
import {
    GeneralFromJSON,
    GeneralFromJSONTyped,
    GeneralToJSON,
    GeneralToJSONTyped,
} from './General';

/**
 * 
 * @export
 * @interface HeadToHeadGeneralRecord
 */
export interface HeadToHeadGeneralRecord {
    /**
     * 
     * @type {General}
     * @memberof HeadToHeadGeneralRecord
     */
    general: General;
    /**
     * 
     * @type {number}
     * @memberof HeadToHeadGeneralRecord
     */
    wins: number;
    /**
     * 
     * @type {number}
     * @memberof HeadToHeadGeneralRecord
     */
    losses: number;
}



/**
 * Check if a given object implements the HeadToHeadGeneralRecord interface.
 */
export function instanceOfHeadToHeadGeneralRecord(value: object): value is HeadToHeadGeneralRecord {
    if (!('general' in value) || value['general'] === undefined) return false;
    if (!('wins' in value) || value['wins'] === undefined) return false;
    if (!('losses' in value) || value['losses'] === undefined) return false;
    return true;
}

export function HeadToHeadGeneralRecordFromJSON(json: any): HeadToHeadGeneralRecord {
    return HeadToHeadGeneralRecordFromJSONTyped(json, false);
}

export function HeadToHeadGeneralRecordFromJSONTyped(json: any, ignoreDiscriminator: boolean): HeadToHeadGeneralRecord {
    if (json == null) {
        return json;
    }
    return {
        
        'general': GeneralFromJSON(json['general']),
        'wins': json['wins'],
        'losses': json['losses'],
    };
}

export function HeadToHeadGeneralRecordToJSON(json: any): HeadToHeadGeneralRecord {
    return HeadToHeadGeneralRecordToJSONTyped(json, false);
}

export function HeadToHeadGeneralRecordToJSONTyped(value?: HeadToHeadGeneralRecord | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'general': GeneralToJSON(value['general']),
        'wins': value['wins'],
        'losses': value['losses'],
    };
}