// src/api/models/HeadToHeadMapRecord.ts
/* tslint:disable */
/* eslint-disable */
/**
 * radarvan
 * Stats for generals
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface HeadToHeadMapRecord
 */
export interface HeadToHeadMapRecord {
    /**
     * 
     * @type {string}
     * @memberof HeadToHeadMapRecord
     */
    map: string;
    /**
     * 
     * @type {number}
     * @memberof HeadToHeadMapRecord
     */
    player1Wins: number;
    /**
     * 
     * @type {number}
     * @memberof HeadToHeadMapRecord
     */
    player2Wins: number;
}

/**
 * Check if a given object implements the HeadToHeadMapRecord interface.
 */
export function instanceOfHeadToHeadMapRecord(value: object): value is HeadToHeadMapRecord {
    if (!('map' in value) || value['map'] === undefined) return false;
    if (!('player1Wins' in value) || value['player1Wins'] === undefined) return false;
    if (!('player2Wins' in value) || value['player2Wins'] === undefined) return false;
    return true;
}

export function HeadToHeadMapRecordFromJSON(json: any): HeadToHeadMapRecord {
    return HeadToHeadMapRecordFromJSONTyped(json, false);
}

export function HeadToHeadMapRecordFromJSONTyped(json: any, ignoreDiscriminator: boolean): HeadToHeadMapRecord {
    if (json == null) {
        return json;
    }
    return {
        
        'map': json['map'],
        'player1Wins': json['player1Wins'],
        'player2Wins': json['player2Wins'],
    };
}

export function HeadToHeadMapRecordToJSON(json: any): HeadToHeadMapRecord {
    return HeadToHeadMapRecordToJSONTyped(json, false);
}

export function HeadToHeadMapRecordToJSONTyped(value?: HeadToHeadMapRecord | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'map': value['map'],
        'player1Wins': value['player1Wins'],
        'player2Wins': value['player2Wins'],
    };
}